bug:-toast issue,new user cred issue, mapper issue

+++ data/tools/sunlight/src/main/java/fit/asta/health/sunlight/remote/model/SessionDetailBody.kt
@@ -8,7 +8,7 @@ data class SessionDetailBody(
     @SerializedName("uid") var uid: String? = null,
     @SerializedName("con") var con: Int? = null,
     @SerializedName("dur") var dur: Long? = null,
-    @SerializedName("temp") var temp: Int? = null,
+    @SerializedName("temp") var temp: Double? = null,
     @SerializedName("uv") var uv: Int? = null,
     @SerializedName("spf") var spf: String? = null,
     @SerializedName("start") var start: String? = null,
